Moved proj.evaluators to mttl.evaluators
